move library to SharedCode folder
