+++ SharedCode/db/radar_table_api.js
