+++ SharedCode/ruten_api.js
