+++ SharedCode/telegram_api.js
