+++ SharedCode/utility.js
